Revert "use long instead of integer in SipApplicationSession"

+++ dev/com.ibm.ws.sipcontainer/src/com/ibm/ws/sip/container/asynch/AsynchronousWorkListenerWrapper.java
@@ -97,7 +97,7 @@ public class AsynchronousWorkListenerWrapper extends RoutedTask{
 	 * 
 	 * @param asynchWorkListener
 	 */
-	public AsynchronousWorkListenerWrapper(AsynchronousWorkListener asynchWorkListener, long queueIndex) {
+	public AsynchronousWorkListenerWrapper(AsynchronousWorkListener asynchWorkListener, int queueIndex) {
 		_appAsynchWorkListener = asynchWorkListener;
 		_index = queueIndex;
 	}
