+++ dev/com.ibm.ws.sipcontainer/src/com/ibm/ws/sip/container/was/ThreadLocalStorage.java
@@ -100,7 +100,7 @@ public class ThreadLocalStorage
 		
 		if (c_logger.isTraceFailureEnabled()) {
 			try {
-				long tuQueueIndex = SipApplicationSessionImpl.extractAppSessionCounter(tu.getApplicationId()) % NativeMessageDispatchingHandler.s_dispatchers;
+				int tuQueueIndex = SipApplicationSessionImpl.extractAppSessionCounter(tu.getApplicationId()) % NativeMessageDispatchingHandler.s_dispatchers;
 				
 				if(ThreadLocalStorage.getQueueId() == null || ! ThreadLocalStorage.getQueueId().equals(tuQueueIndex)) {
 					c_logger.traceFailure(tu, "traceThreadModelValidity", "Threading model violation! " +
